[Add] Added battle field
