--- Assets/Resources/Scriptable/Cutscene/Tutorial.asset
@@ -1,25 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: Tutorial
-  m_EditorClassIdentifier: Assembly-CSharp::Cutscene
-  playableAsset: {fileID: 11400000, guid: 8a6980645c67d1346aa628fd1b1cf62c, type: 2}
-  dialogueSkipIndex: 06000000
-  cutsceneSkipTime:
-  - 26.7
-  - 28.5
-  cutsceneSkipSound:
-  - Town
-  cutsceneParticipants:
-  - Zio
-  - Pitch
-  - Timmy
