+++ Assets/Scenes/Field.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 0c3bc17f77065814e8e6b3764ff93783
-NativeFormatImporter:
+guid: c6a2d424ba2d3f745b3e05e1626f3c6f
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
