+++ Assets/Scenes/Top View.unity
@@ -482,7 +482,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &511660986
 RectTransform:
   m_ObjectHideFlags: 0
