+++ Assets/Script/General/Manager/CutsceneManager.cs
@@ -211,12 +211,4 @@ public class CutsceneManager : MonoBehaviour
     }
 }
 
-[CreateAssetMenu (menuName = "Scriptable Object/Cutscene", fileName = "Cutscene", order = int.MaxValue - 1)]
-public class Cutscene : ScriptableObject
-{
-    [SerializeField] public PlayableAsset playableAsset;
-    [SerializeField] public int[] dialogueSkipIndex;
-    [SerializeField] public double[] cutsceneSkipTime;
-    [SerializeField] public string[] cutsceneSkipSound;
-    [SerializeField] public string[] cutsceneParticipants;
-}
+
