+++ Assets/Script/TopView/Cutscene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa0f61023e4c65044a178f4c22bbc27e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
